feat(button): abstract base class
