+++ elements/pfe-button/test/pfe-button.spec.ts
@@ -67,12 +67,4 @@ describe('<pfe-button>', function() {
 
     expect(lightDomBtn).dom.to.equal('<button>Button</button>');
   });
-
-  it('should send a pfe-button:click event on click', async function() {
-    const el = await createFixture<PfeButton>(element);
-    setTimeout(() => el.querySelector('button')!.click());
-    const event = await oneEvent(document, 'pfe-button:click');
-    expect(event).to.be.ok;
-    expect(event.type).to.equal('pfe-button:click');
-  });
 });
